Cleanup 3: cli.js, helpers.js & their impl
